+++ RelativeTimeChecker.js
@@ -5,7 +5,8 @@ function setDatetime(datetime, element, setAttribute = true) {
     let attribute;
     if (datetime === null) {
         attribute = null;
-        element.removeAttribute('datetime');
+        if (setAttribute)
+            element.removeAttribute('datetime');
     }
     else if (datetime instanceof Temporal.ZonedDateTime) {
         attribute = (new Date(datetime.epochMilliseconds)).toISOString();