+++ tests.js
@@ -3,6 +3,11 @@ import {Datetime_global} from "./Datetime_global.js";
 
 // get the current time
 const now = new Datetime_global('2025-07-21T23:59:00', 'Europe/Amsterdam');
-console.log(now.toString(), now.toJSON(), now.useOldJSON);
-now.useOldJSON = 1;
-console.log(now.toString(), now.toJSON(), now.useOldJSON);
+console.log(now.toString());
+const plain = now.toTemporalZonedDateTime().toPlainDateTime();
+console.log(now.templateFormat`PlainDateTime: ${plain}
+
+Datetime_global: ${Datetime_global}
+Date.now: ${Date.now}
+Date.UTC: ${Date.UTC}
+Date: ${Date}`);